Moved prefix of AuthController path to @RequestMapping
